+++ src/simulation/utility.h
@@ -5,6 +5,6 @@
 namespace LGen {
 	class Utility final {
 	public:
-		static float utility(const LRender::ReportAgent &report);
+		static double utility(const LRender::ReportAgent &report);
 	};
 }
